+++ src/main/kotlin/codegen/FlutterXPluginInfo.kt
@@ -1,11 +1,11 @@
 package codegen
 // 自动生成的插件信息类,不要修改这个文件,否则会导致插件功能失效
 object FlutterXPluginInfo {
-    const val VERSION: String = "5.6.1.251"
+    const val VERSION: String = "5.6.2.243"
     const val CHANGELOG: String = """
-<h2>5.6.1 - 2025-05-06</h2>
+<h2>5.6.2 - 2025-05-08</h2>
 
-<ul><li>Unused package logic detection optimization <a href="https://github.com/mdddj/dd_flutter_idea_plugin/issues/78">#78</a></li><li><a href="https://github.com/mdddj/dd_flutter_idea_plugin/blob/master/CHANGELOG.md">full changelog</a></li></ul>
+<ul><li>新增 dio重启逻辑</li><li>优化 l10n编辑面板，新增 preview text 选项，在 tree中预览文本</li></ul>
 """
 
 }